Write a new stream, CheckpointStream
CheckpointStream is designed to be simpler to use than a bunch of Deques
for maintaining a stream that can be undone (for example, if part of a
sequential parse fails). It is also meant to be more efficient, since it
stores only characters that are read rather than copying over the entire
stream.