Add strong checkpoints
Strong checkpoints exist so that way NegativeLookaheadParser can always
reset the stream after it performs its lookahead check. This wasn't
occuring when the lookahead is complex (for example, a Sequence), so
strong checkpoints were devised to fix this.

Strong checkpoints always reset the stream to the state that the stream
was in when the strong checkpoint was created. They aren't tested, nor
are they particularly well defined, but they seem to work with the current
grammar that I'm using.