Change how tokens handle internal newlines
When splitting a token over multiple lines, the token will always return
itself if it doesn't contain an embedded newline, or if the token itself
is a newline.

However, if the token has a newline, and is more than just a newline, it
will make sure to include newlines for all of its subtokens but the last
one. This avoids introducing unnecessary newlines in some circumstances.